+++ Final_BackEnd/src/routes/compartir.route.ts
@@ -1,12 +1,14 @@
 import express from 'express';
 import CompartirRecetasController from '../controller/compartir.controller.js';
+import authMiddleware from '../middleware/authMiddleware.js';
 
 const router = express.Router();
 const compartirController = new CompartirRecetasController();
 
 // Compartir una receta
-router.post('/compartir', async (req, res) => {
+router.post('/compartir', authMiddleware, async (req, res) => {
   const { id_usuario, id_receta, id_region, id_tipo } = req.body;
+  
   try {
     const compartirId = await compartirController.compartirReceta({ id_usuario, id_receta, id_region, id_tipo });
     res.status(201).json({ message: 'Receta compartida con éxito', compartirId });
@@ -38,7 +40,7 @@ router.get('/compartidas', async (req, res) => {
 });
 
 // Obtener las recetas compartidas por un usuario
-router.get('/usuario/:id', async (req, res) => {
+router.get('/usuario/:id', authMiddleware, async (req, res) => {
   const { id } = req.params;
   try {
     const recetasCompartidas = await compartirController.getCompartidasByUsuario(Number(id));
@@ -55,9 +57,19 @@ router.get('/usuario/:id', async (req, res) => {
 });
 
 // Eliminar una receta compartida
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', authMiddleware, async (req, res) => {
   const { id } = req.params;
+  const userId = (req as any).userId;
   try {
+    const recetaExistente = await compartirController.getCompartidasById(Number(id));
+
+    if (!recetaExistente) {
+      return res.status(404).json({ message: 'Receta no encontrada' });
+    }
+
+    if (recetaExistente.id_usuario !== userId) {
+      return res.status(403).json({ message: 'No tienes permiso para eliminar esta calificación' });
+    }
     const deletedRows = await compartirController.deleteCompartir(Number(id));
     if (deletedRows === 0) {
       res.status(404).json({ message: 'Receta compartida no encontrada' });
